requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.1.0
python-dotenv==1.0.0
requests==2.28.2
stripe==5.0.0
urllib3==1.26.14
